+++ boss-rally/BRD3D/split-win32/meth_unk00334/meth_unk00334_GetDirectDrawGUID.c
@@ -1,11 +1,11 @@
 #include "types-win32.h"
 //----- (1001B960) --------------------------------------------------------
 GUID *__thiscall meth_unk00334_GetDirectDrawGUID(unk00334 *this) {
-  GUID *result; // eax
+    GUID *result; // eax
 
-  if (((int)this->ddCapsFlags & 2) != 0)
-    result = 0;
-  else
-    result = (GUID *)&this->lpGUID;
-  return result;
+    if ((this->ddCapsFlags & 2) != 0)
+        result = 0;
+    else
+        result = (GUID *)&this->lpGUID;
+    return result;
 }
